// TSIC-Core-Angular/src/frontend/tsic-app/src/app/core/api/models/models/ClubTeamManagementDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ClubTeamManagementDto = {
    clubTeamId: number | string;
    clubTeamName: string;
    clubTeamGradYear: string;
    clubTeamLevelOfPlay: string;
    isActive: boolean;
    hasBeenUsed: boolean;
    hasBeenRegisteredForAnyEvent: boolean;
};